server_python/requirements.txt
# FastAPI and ASGI server
fastapi>=0.115.0
uvicorn[standard]>=0.30.0

# Data validation
pydantic>=2.9.0
pydantic-settings>=2.5.0

# Authentication
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.9

# AWS SDK
boto3>=1.35.0

# HTTP client
httpx>=0.27.0

# Cryptography
cryptography>=43.0.0

# Environment variables
python-dotenv>=1.0.0

# Testing
pytest>=8.3.0
pytest-asyncio>=0.23.0
httpx>=0.27.0
